finance-assistant/backend/app/requirements.txt
fastapi
uvicorn
sqlalchemy
pymssql
pydantic
python-jose
passlib
python-multipart
langchain
langchain-core
langchain-community
langchain-openai
langchain-experimental
langgraph
pypdf
pdfminer.six
azure-identity
azure-storage-blob
